"Money" fields converted to decimal (closes #2)

+++ OneSky.CSharp/OneSky.CSharp/Json/Objects/IQuotationDetails.cs
@@ -4,9 +4,9 @@
 
     public interface IQuotationDetails
     {
-        float PerWordCost { get; }
+        decimal PerWordCost { get; }
 
-        float TotalCost { get; }
+        decimal TotalCost { get; }
 
         DateTime WillCompleteAt { get; }
